Fix installer for linux
