+++ .gitignore
@@ -1,6 +1,7 @@
 /bin/
 /sweetiebot/override.go
 /sweetiebot_data.sql
+/sweetiebot_transcripts.sql
 /transcripts/
 *.md5
 *.test
