+++ markovmodule/MarkovModule.go
@@ -95,8 +95,8 @@ func (c *episodeGenCommand) Usage(info *bot.GuildInfo) *bot.CommandUsage {
 	return &bot.CommandUsage{
 		Desc: "Randomly generates a my little pony episode using a markov chain, up to a maximum line count of `lines`. Will be sent via PM if the line count exceeds 5.",
 		Params: []bot.CommandUsageParam{
-			{Name: "lines", Desc: "Number of dialogue lines to generate", Optional: false},
-			{Name: "single", Desc: "The markov chain uses double-lookback by default, if this is specified, will revert to single-lookback, which produces much more chaotic results.", Optional: false},
+			{Name: "lines", Desc: "Number of dialogue lines to generate", Optional: true},
+			{Name: "single", Desc: "The markov chain uses double-lookback by default, if this is specified, will revert to single-lookback, which produces much more chaotic results.", Optional: true},
 		},
 	}
 }
@@ -128,7 +128,7 @@ func (c *episodeQuoteCommand) Process(args []string, msg *discordgo.Message, ind
 		arg := strings.ToLower(args[0])
 		switch arg {
 		case "action":
-			lines = []bot.Transcript{info.Bot.DB.GetCharacterQuote("ACTION")}
+			lines = []bot.Transcript{info.Bot.DB.GetCharacterQuote("")}
 		case "speech":
 			lines = []bot.Transcript{info.Bot.DB.GetSpeechQuote()}
 		default:
@@ -162,7 +162,7 @@ func (c *episodeQuoteCommand) Process(args []string, msg *discordgo.Message, ind
 	process := make([]string, 0, len(lines))
 	for _, v := range lines {
 		l := ""
-		if v.Speaker == "ACTION" {
+		if v.Speaker == "" {
 			if v.Text != "" {
 				l = "[" + v.Text + "]"
 			}
@@ -251,120 +251,3 @@ func (c *shipCommand) Usage(info *bot.GuildInfo) *bot.CommandUsage {
 		},
 	}
 }
-
-/*
-func (sb *SweetieBot) ingestEpisode(file string, season int, episode int) {
-	f, err := ioutil.ReadFile(file)
-	if err != nil {
-		fmt.Println(err.Error())
-	}
-	s := strings.Split(strings.Replace(string(f), "\r", "", -1), "\n")
-
-	songmode := false
-	lastcharacter := ""
-	adjust := 0
-	for i := 0; i < len(s); i++ {
-		if len(s[i]) > 0 {
-			if s[i][0] == '[' {
-				action := s[i][1 : len(s[i])-1]
-				sb.DB.AddTranscript(season, episode, i-adjust, "ACTION", action)
-				if !songmode {
-					lastcharacter = action
-				}
-			} else {
-				split := strings.SplitN(s[i], ":", 2)
-				songmode = (len(split) < 2)
-				if songmode {
-					prev := sb.DB.GetTranscript(season, episode, i-1-adjust, i-1-adjust)
-					if len(prev) != 1 {
-						fmt.Println(season, " ", episode, " ", i-adjust)
-						return
-					}
-					if prev[0].Speaker == "ACTION" && prev[0].Text == lastcharacter {
-						adjust++
-						sb.DB.RemoveTranscript(season, episode, i-adjust)
-					}
-					sb.DB.AddTranscript(season, episode, i-adjust, lastcharacter, strings.TrimSpace(split[0]))
-				} else {
-					lastcharacter = strings.TrimSpace(split[0])
-					sb.DB.AddTranscript(season, episode, i-adjust, lastcharacter, strings.TrimSpace(split[1]))
-				}
-			}
-		} else {
-			sb.DB.AddTranscript(season, episode, i-adjust, "ACTION", "")
-		}
-	}
-}
-
-func splitSpeaker(speaker string) []string {
-	speakers := strings.Split(strings.Replace(speaker, ", and", " and", -1), " and ")
-	speakers = append(strings.Split(speakers[0], ","), speakers[1:]...)
-	for i, s := range speakers {
-		speakers[i] = strings.Trim(strings.TrimSpace(strings.Replace(s, "Young", "", -1)), "\"")
-	}
-	return speakers
-}
-
-func (sb *SweetieBot) buildMarkov(seasonStart int, episodeStart int) {
-	regex := regexp.MustCompile("[^~!@#$%^&*()_+`=[\\];,./<>?\" \n\r\f\t\v]+[?!.]?")
-
-	sb.DB.sqlResetMarkov.Exec()
-
-	var cur uint64
-	var prev uint64
-	var prev2 uint64
-	for season := seasonStart; season <= 5; season++ {
-		for episode := episodeStart; episode <= 26; episode++ {
-			fmt.Println("Begin Episode", episode, "Season", season)
-			prev = 0
-			prev2 = 0
-			lines := sb.DB.GetTranscript(season, episode, 0, 999999)
-			//lines := []Transcript{ {1, 1, 1, "Twilight", "Twilight went to the bakery to buy some cakes."}, {1, 1, 1, "Twilight", "Twilight went to the library to buy some books"} }
-			fmt.Println("Got", len(lines), "lines")
-
-			for i := range lines {
-				if len(lines[i].Text) == 0 {
-					if lines[i].Speaker != "ACTION" {
-						fmt.Println("UNKNOWN SPEAKER: ", lines[i].Speaker)
-					}
-					cur = sb.DB.AddMarkov(prev, prev2, lines[i].Speaker, "")
-					prev2 = 0
-					prev = cur // Cur will always be 0 here.
-					continue
-				}
-				words := regex.FindAllString(lines[i].Text, -1)
-				speakers := splitSpeaker(lines[i].Speaker)
-				for _, speaker := range speakers {
-					if len(speaker) == 0 {
-						fmt.Println("EMPTY SPEAKER GENERATED FROM \""+lines[i].Speaker+"\" ON LINE: ", lines[i].Text)
-						fmt.Println(speakers)
-					}
-					for j := range words {
-						l := len(words[j])
-						ch := words[j][l-1]
-						switch ch {
-						case '.', '!', '?':
-							words[j] = words[j][:l-1]
-						}
-						if sb.DB.GetMarkovWord(speaker, words[j]) != words[j] {
-							words[j] = strings.ToLower(words[j])
-						}
-						//fmt.Println("AddMarkov: ", prev, prev2, speaker, words[j])
-						cur = sb.DB.AddMarkov(prev, prev2, speaker, words[j])
-						prev2 = prev
-						prev = cur
-
-						switch ch {
-						case '.', '!', '?':
-							//fmt.Println("AddMarkov: ", prev, prev2, speaker, string(ch))
-							cur = sb.DB.AddMarkov(prev, prev2, speaker, string(ch))
-							prev2 = 0
-							prev = 0
-							//prev = sb.DB.AddMarkov(prev, "ACTION", "")
-						}
-					}
-				}
-			}
-		}
-	}
-}*/
