+++ miscmodule/SearchCommand.go
